+++ noodles-bam/src/record/codec/encoder/data/field/value.rs
@@ -1,14 +1,14 @@
 pub mod array;
+mod hex;
 
 use std::io;
 
 use noodles_sam::alignment::record::data::field::Value;
 
+use self::{array::write_array, hex::write_hex};
 use crate::record::codec::encoder::num::{
     write_f32_le, write_i16_le, write_i32_le, write_i8, write_u16_le, write_u32_le, write_u8,
 };
-
-use self::array::write_array;
 
 pub fn write_value(dst: &mut Vec<u8>, value: &Value) -> io::Result<()> {
     match value {
@@ -21,7 +21,7 @@ pub fn write_value(dst: &mut Vec<u8>, value: &Value) -> io::Result<()> {
         Value::UInt32(n) => write_u32_le(dst, *n),
         Value::Float(n) => write_f32_le(dst, *n),
         Value::String(s) => write_string(dst, s),
-        Value::Hex(s) => write_string(dst, s),
+        Value::Hex(s) => write_hex(dst, s)?,
         Value::Array(array) => write_array(dst, array)?,
     }